--- pages/EditDashboard/[dashboard_id]/post.vue
@@ -1,9 +1,0 @@
-<script setup lang="ts">
-
-
-</script>
-
-<template lang="pug">
-button 
-    a pranav
-</template>
