+++ pages/edit-dashboard/[dashboard_id]/post.vue
@@ -1,0 +1,9 @@
+<script setup lang="ts">
+
+
+</script>
+
+<template lang="pug">
+button 
+    a pranav
+</template>
